# backend/openapi.yml
openapi: 3.1.0
info:
  version: 1.0.0
  title: HireSense (知聘) - Backend
  description: HireSense 后端接口，涵盖了用户鉴权、档案管理、实时沉浸式面试交互以及多维度面试报告生成等核心能力。
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发环境的服务地址
security:
  - bearerAuth: []
...
paths:
...
  /api/interview/start:
    post:
      tags: [Interview]
      summary: 开始一场面试
      description: 基于用户当前档案数据生成并启动一个新的 AI 面试会话，并获取首个提问。
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InterviewStartRequest" }
      responses:
        "200":
          description: 面试正式启动并返回该场面试 ID 及首个问题
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InterviewStartResponse" }
  /api/interview/stop:
    post:
      tags: [Interview]
      summary: 结束一场面试
      description: 候选人主动放弃或时间耗尽时调用，停止计费并中止本次会话进程。（通常在系统判断回答 ending=true 时无需显式调用）
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InterviewStopRequest" }
      responses:
        "200":
          description: 成功阻断当前面试场次
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InterviewStopResponse" }
  /api/interview/reply:
    post:
      tags: [Interview]
      summary: 生成面试官回复
      description: 提交候选人录制的音频流或者直接提交文本，AI面试官会进行意图分析并进行下一轮追问。
      parameters:
        - name: id
          in: query
          required: true
          description: 当前进行的面试场次唯一 ID
          schema:
            type: string
            format: uuid
            examples:
              - "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InterviewReplyRequest" }
          audio/wav:
            schema:
              description: 直接上传用户作答时的音频流
              type: string
              format: binary
      responses:
        "200":
          description: AI 已生成针对本次问答的反馈与下个问题
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InterviewReplyResponse" }
  /api/interview/tts:
    post:
      tags: [Interview]
      summary: 生成面试官语音
      description: 将前端已经收到的面试官文本合成为 mp3 语音，用于在面试页自动播放。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InterviewTTSRequest" }
      responses:
        "200":
          description: 成功生成面试官语音
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "传递形式: `Authorization: Bearer <Your_Token>`"
  schemas:
...
    InterviewMode:
      type: string
      description: 面试模式
      enum: ["single", "panel_trio"]
      examples:
        - "panel_trio"
    InterviewSpeakerRole:
      type: string
      description: 发言的 AI 面试官角色
      enum: ["ai", "hr", "tech_lead", "executive"]
      examples:
        - "tech_lead"
    InterviewStartRequest:
      type: object
      properties:
        mode:
          $ref: "#/components/schemas/InterviewMode"
    InterviewStartResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "AI 面试官已就绪"
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
              description: 新创建的本次面试 ID
              examples:
                - "550e8400-e29b-41d4-a716-446655440000"
            reply:
              type: string
              description: AI 面试官返回的首个问题
              examples:
                - "先做一个简短的自我介绍，并重点说明你最近一段项目经历里最有代表性的技术挑战。"
            mode:
              $ref: "#/components/schemas/InterviewMode"
            speaker_role:
              $ref: "#/components/schemas/InterviewSpeakerRole"
          required: [id, reply, mode, speaker_role]
      required: [success, message, data]
    InterviewStopRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: 需要终止的面试 ID
          examples:
            - "550e8400-e29b-41d4-a716-446655440000"
      required: [id]
    InterviewStopResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "已被强制标记为停止"
        data: {}
      required: [success, message, data]
    InterviewReplyRequest:
      type: object
      description: 候选人的面试回答
      properties:
        text:
          type: string
          description: 针对上一轮面试官提问的用户答案
          examples:
            - "关于 React 的优化，我平时会经常使用 React.memo 和 useMemo 来防止组件的非必要渲染..."
      required: [text]
    InterviewTTSRequest:
      type: object
      description: 需要合成为面试官语音的文本内容
      properties:
        text:
          type: string
          description: 前端已收到并准备播放的面试官文本
          examples:
            - "你好，请先做一个简短的自我介绍。"
      required: [text]
    InterviewReplyResponse:
      type: object
      properties:
        success:
          type: boolean
          examples:
            - true
        message:
          type: string
          examples:
            - "交互成功"
        data:
          type: object
          properties:
            reply:
              type: string
              description: AI 面试官作出的回应以及下一轮追问
              examples:
                - "思路清晰！既然提到了 useMemo，你能顺便讲讲什么情况下过度滥用会导致反效果吗？"
            ending:
              type: boolean
              description: 面试是否结束
              examples:
                - false
            mode:
              $ref: "#/components/schemas/InterviewMode"
            speaker_role:
              $ref: "#/components/schemas/InterviewSpeakerRole"
          required: [reply, ending, mode, speaker_role]
      required: [success, message, data]
...